pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "bountyhound-local"
version = "1.0.0"
description = "Autonomous bug bounty hunting swarm powered by local LLMs on H100 GPU"
requires-python = ">=3.10"
dependencies = [
    "vllm>=0.6.0",
    "openai>=1.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "flower>=2.0.0",
    "aiosqlite>=0.20.0",
    "playwright>=1.45.0",
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "APScheduler>=3.10.0",
]

[project.scripts]
bhl = "cli:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
